backend/requirements.txt
fastapi[standard]
python-dotenv 
google-genai
sqlalchemy
psycopg2-binary
alembic
pydantic-settings